fix issue of sometime adding duplicate tries on ENTER
Sometimes, on pressing ENTER key in text box, the duplicate checking
is not working. This is becase, the word is added to the list only after
a successful network call completion.  If there is more network delay,
subsequent checkDuplicate returns true and further queues up the calls.

To fix it, separate data structure is kept to check for duplicates alone.
Any word that has been already processed or in processing in network, will
be checked against it and udpated immediately.